refactor(6/introduce): ThirdInteraction 컴포넌트 분리
