+++ src/app/6/introduce/_components/ThirdInteraction.tsx
@@ -1,12 +1,12 @@
 'use client';
 
-import React, { useLayoutEffect, useRef, useState, useEffect } from 'react';
+import React, { useLayoutEffect, useRef } from 'react';
 import gsap from 'gsap';
-import { DotLottiePlayer as Player } from '@dotlottie/react-player';
-import '@dotlottie/react-player/dist/index.css';
-import Image from 'next/image';
-import { FaGithub } from 'react-icons/fa';
-import { STAFF_DATA, StaffMember } from '../_constants/staff.ts';
+import { STAFF_DATA } from '../_constants/staff.ts';
+import MarqueeItem from './MarqueeItem.tsx';
+import TechStack from './TechStack.tsx';
+import CountdownTimer from './CountdownTimer.tsx';
+
 
 interface ThirdInteractionProps {
   refs: {
@@ -19,157 +19,6 @@ interface ThirdInteractionProps {
     applySectionRef: React.RefObject<HTMLDivElement>;
   };
 }
-
-const RightArrow = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="w-8 h-8 md:w-12 md:h-12 mx-4 text-white"
-  >
-    <path d="M5 12h14" />
-    <path d="m12 5 7 7-7 7" />
-  </svg>
-);
-
-const LeftArrow = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="w-8 h-8 md:w-12 md:h-12 mx-4 text-white"
-  >
-    <path d="M19 12H5" />
-    <path d="m12 19-7-7 7-7" />
-  </svg>
-);
-
-const MarqueeItem = ({ text, direction }: { text: string; direction: 'left' | 'right' }) => (
-  <div className="flex items-center flex-shrink-0">
-    <span className="text-2xl md:text-4xl font-semibold text-white mx-4">{text}</span>
-    {direction === 'right' ? <RightArrow /> : <LeftArrow />}
-  </div>
-);
-
-const TechStack = ({
-  lottie,
-  name,
-  fRef,
-  staff,
-}: {
-  lottie: string;
-  name: string;
-  fRef: React.RefObject<HTMLDivElement>;
-  staff: StaffMember[];
-}) => (
-  <div ref={fRef} className="absolute inset-0 flex items-center justify-center p-4">
-    <div className="flex flex-col items-center justify-center w-full max-w-5xl">
-      <div className="flex items-center justify-center">
-        <div className="w-24 h-24 sm:w-32 sm:h-32 md:w-40 md:h-40">
-          <Player src={lottie} autoplay loop style={{ width: '100%', height: '100%' }} />
-        </div>
-        <span className="text-3xl sm:text-5xl md:text-6xl font-bold text-black ml-2 md:ml-6">{name}</span>
-      </div>
-      <div
-        className={
-          name === 'PM & Design'
-            ? 'grid grid-cols-2 gap-x-8 gap-y-6 mt-6 md:mt-10'
-            : 'flex flex-wrap justify-center gap-4 md:gap-6 mt-6 md:mt-10'
-        }
-      >
-        {staff.map(member => (
-          <div key={member.name} className="flex flex-col items-center text-center w-24 sm:w-28">
-            <div className="w-20 h-20 sm:w-24 sm:h-24 rounded-full overflow-hidden mb-2">
-              <Image src={member.imageUrl} alt={member.name} width={128} height={128} className="object-cover w-full h-full" />
-            </div>
-            <div className="flex items-center space-x-1.5">
-              <p className="font-bold text-sm sm:text-base text-black whitespace-nowrap">{member.name}</p>
-              {member.github && (
-                <a href={member.github} target="_blank" rel="noopener noreferrer">
-                  <FaGithub className="w-3.5 h-3.5 sm:w-4 sm:h-4 text-gray-700 hover:text-black" />
-                </a>
-              )}
-            </div>
-            <p className="text-gray-600 text-xs sm:text-sm">
-              {member.position}
-              {name !== 'Management' && ' 파트장'}
-            </p>
-          </div>
-        ))}
-      </div>
-    </div>
-  </div>
-);
-
-const CountdownTimer = () => {
-  const [isClient, setIsClient] = useState(false);
-
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
-
-  const calculateTimeLeft = () => {
-    const difference = +new Date('2025-09-05T23:59:59+09:00') - +new Date();
-    let timeLeft: { days?: number; hours?: number; minutes?: number; seconds?: number } = {};
-
-    if (difference > 0) {
-      timeLeft = {
-        days: Math.floor(difference / (1000 * 60 * 60 * 24)),
-        hours: Math.floor((difference / (1000 * 60 * 60)) % 24),
-        minutes: Math.floor((difference / 1000 / 60) % 60),
-        seconds: Math.floor((difference / 1000) % 60),
-      };
-    }
-    return timeLeft;
-  };
-
-  const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
-
-  useEffect(() => {
-    if (!isClient) return;
-
-    const timer = setTimeout(() => {
-      setTimeLeft(calculateTimeLeft());
-    }, 1000);
-
-    return () => clearTimeout(timer);
-  });
-
-  const hasTimeLeft = Object.values(timeLeft).some(value => value > 0);
-
-  if (!isClient) {
-    return (
-      <div className="text-center my-4">
-        <div className="text-5xl md:text-7xl font-bold text-black font-mono" style={{ opacity: 0 }}>
-          00:00:00:00
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="text-center my-4">
-      {hasTimeLeft ? (
-        <div className="text-5xl md:text-7xl font-bold text-black font-mono">
-          <span>{String(timeLeft.days).padStart(2, '0')}:</span>
-          <span>{String(timeLeft.hours).padStart(2, '0')}:</span>
-          <span>{String(timeLeft.minutes).padStart(2, '0')}:</span>
-          <span>{String(timeLeft.seconds).padStart(2, '0')}</span>
-        </div>
-      ) : (
-        <span className="text-2xl md:text-4xl font-bold text-black">모집이 마감되었습니다.</span>
-      )}
-    </div>
-  );
-};
 
 const ThirdInteraction = ({ refs }: ThirdInteractionProps) => {
   const { sectionRef, managementGroupRef, androidGroupRef, webGroupRef, serverGroupRef, designGroupRef, applySectionRef } =
